Move `Themes` to `Preferences` Component
